fix[NO TICKET] separate messageBridge declaration
