+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayActivity.kt
@@ -23,22 +23,13 @@ import com.rakuten.tech.mobile.miniapp.MiniApp
 import com.rakuten.tech.mobile.miniapp.MiniAppInfo
 import com.rakuten.tech.mobile.miniapp.MiniAppSdkConfig
 import com.rakuten.tech.mobile.miniapp.ads.AdMobDisplayer
-import com.rakuten.tech.mobile.miniapp.errors.MiniAppAccessTokenError
-import com.rakuten.tech.mobile.miniapp.errors.MiniAppPointsError
 import com.rakuten.tech.mobile.miniapp.file.MiniAppCameraPermissionDispatcher
 import com.rakuten.tech.mobile.miniapp.file.MiniAppFileChooserDefault
 import com.rakuten.tech.mobile.miniapp.file.MiniAppFileDownloaderDefault
-import com.rakuten.tech.mobile.miniapp.js.MessageToContact
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import com.rakuten.tech.mobile.miniapp.js.NativeEventType
-import com.rakuten.tech.mobile.miniapp.js.chat.ChatBridgeDispatcher
-import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
-import com.rakuten.tech.mobile.miniapp.js.userinfo.Points
-import com.rakuten.tech.mobile.miniapp.js.userinfo.TokenData
-import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridgeDispatcher
 import com.rakuten.tech.mobile.miniapp.navigator.ExternalResultHandler
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
-import com.rakuten.tech.mobile.miniapp.permission.AccessTokenScope
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppDevicePermissionType
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.MiniAppDisplayActivityBinding
@@ -54,6 +45,7 @@ class MiniAppDisplayActivity : BaseActivity(), PreloadMiniAppWindow.PreloadMiniA
     override val pageName: String = this::class.simpleName ?: ""
     override val siteSection: String = this::class.simpleName ?: ""
 
+    private lateinit var miniAppMessageBridge: MiniAppMessageBridge
     private lateinit var miniAppNavigator: MiniAppNavigator
     private var miniappPermissionCallback: (isGranted: Boolean) -> Unit = {}
     private var miniappCameraPermissionCallback: (isGranted: Boolean) -> Unit = {}
@@ -101,175 +93,6 @@ class MiniAppDisplayActivity : BaseActivity(), PreloadMiniAppWindow.PreloadMiniA
 
     private var appInfo: MiniAppInfo? = null
     private var appUrl: String? = null
-
-    private fun hasMessage(message: String): Boolean {
-        return message.isNotEmpty()
-    }
-
-
-    private var miniAppMessageBridge: MiniAppMessageBridge = object : MiniAppMessageBridge() {
-
-        override fun getUniqueId(
-            onSuccess: (uniqueId: String) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            AppSettings.instance.uniqueIdError.let {
-                if (hasMessage(it)) {
-                    onError(it)
-                    return
-                }
-            }
-            onSuccess(AppSettings.instance.uniqueId)
-        }
-
-        override fun getMessagingUniqueId(
-            onSuccess: (uniqueId: String) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            AppSettings.instance.messagingUniqueIdError.let {
-                if (hasMessage(it)) {
-                    onError(it)
-                    return
-                }
-            }
-            onSuccess("TEST-MESSAGE_UNIQUE-ID-01234")
-        }
-
-        override fun getMauid(
-            onSuccess: (mauid: String) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            AppSettings.instance.mauIdError.let {
-                if (hasMessage(it)) {
-                    onError(it)
-                    return
-                }
-            }
-            onSuccess("TEST-MAUID-01234-56789")
-        }
-
-        override fun requestDevicePermission(
-            miniAppPermissionType: MiniAppDevicePermissionType,
-            callback: (isGranted: Boolean) -> Unit
-        ) {
-            miniappPermissionCallback = callback
-            ActivityCompat.requestPermissions(
-                this@MiniAppDisplayActivity,
-                AppPermission.getDevicePermissionRequest(miniAppPermissionType),
-                AppPermission.getDeviceRequestCode(miniAppPermissionType)
-            )
-        }
-
-        override fun sendJsonToHostApp(
-            jsonStr: String,
-            onSuccess: (String) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            jsonStr.let {
-                val message: String
-                if (it.isNotBlank()) {
-                    message = it
-                    onSuccess(message)
-                } else {
-                    message = getString(R.string.error_send_json_to_host_app_please_try_again)
-                    onError(message)
-                }
-                if (isAvailable) {
-                    showToastMessage(message)
-                }
-            }
-        }
-    }
-
-    private val userInfoBridgeDispatcher = object : UserInfoBridgeDispatcher {
-
-        override fun getUserName(
-            onSuccess: (userName: String) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            AppSettings.instance.profileName.let {
-                if (hasMessage(it)) {
-                    onSuccess(it)
-                    return
-                }
-            }
-            onError("User name is not found.")
-        }
-
-        override fun getProfilePhoto(
-            onSuccess: (profilePhoto: String) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            AppSettings.instance.profilePictureUrlBase64.let {
-                if (hasMessage(it)) {
-                    onSuccess(it)
-                    return
-                }
-            }
-            onError("Profile photo is not found.")
-        }
-
-        override fun getAccessToken(
-            miniAppId: String,
-            accessTokenScope: AccessTokenScope,
-            onSuccess: (tokenData: TokenData) -> Unit,
-            onError: (tokenError: MiniAppAccessTokenError) -> Unit
-        ) {
-            if (AppSettings.instance.accessTokenError != null) {
-                onError(AppSettings.instance.accessTokenError!!)
-            } else {
-                onSuccess(AppSettings.instance.tokenData)
-            }
-        }
-
-        override fun getContacts(
-            onSuccess: (contacts: ArrayList<Contact>) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            if (AppSettings.instance.isContactsSaved)
-                onSuccess(AppSettings.instance.contacts)
-            else
-                onError("There is no contact found in HostApp.")
-        }
-
-        override fun getPoints(
-            onSuccess: (points: Points) -> Unit,
-            onError: (pointsError: MiniAppPointsError) -> Unit
-        ) {
-            val points = AppSettings.instance.points
-            if (points != null) onSuccess(points)
-            else onError(MiniAppPointsError.custom("There is no points found in HostApp."))
-        }
-    }
-
-    val chatWindow = ChatWindow(this@MiniAppDisplayActivity)
-    val chatBridgeDispatcher = object : ChatBridgeDispatcher {
-
-        override fun sendMessageToContact(
-            message: MessageToContact,
-            onSuccess: (contactId: String?) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            chatWindow.openSingleContactSelection(message, onSuccess, onError)
-        }
-
-        override fun sendMessageToContactId(
-            contactId: String,
-            message: MessageToContact,
-            onSuccess: (contactId: String?) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            chatWindow.openSpecificContactIdSelection(contactId, message, onSuccess, onError)
-        }
-
-        override fun sendMessageToMultipleContacts(
-            message: MessageToContact,
-            onSuccess: (contactIds: List<String>?) -> Unit,
-            onError: (message: String) -> Unit
-        ) {
-            chatWindow.openMultipleContactSelections(message, onSuccess, onError)
-        }
-    }
 
     companion object {
         private const val appIdTag = "app_id_tag"
@@ -484,14 +307,20 @@ class MiniAppDisplayActivity : BaseActivity(), PreloadMiniAppWindow.PreloadMiniA
 
     @Suppress("OverridingDeprecatedMember")
     private fun setupMiniAppMessageBridge() {
+        miniAppMessageBridge = getMessageBridge(this) { onDevicePermissionResultCallback ->
+            miniappPermissionCallback = onDevicePermissionResultCallback
+        }
         miniAppMessageBridge.setAdMobDisplayer(AdMobDisplayer(this@MiniAppDisplayActivity))
         miniAppMessageBridge.allowScreenOrientation(true)
 
+        // setup UserInfoBridgeDispatcher
+        val userInfoBridgeDispatcher = getUserInfoBridgeDispatcher()
         miniAppMessageBridge.setUserInfoBridgeDispatcher(userInfoBridgeDispatcher)
 
         // setup ChatBridgeDispatcher
+        val chatWindow = ChatWindow(this@MiniAppDisplayActivity)
+        val chatBridgeDispatcher = getChatBridgeDispatcher(chatWindow)
         miniAppMessageBridge.setChatBridgeDispatcher(chatBridgeDispatcher)
-
         miniAppMessageBridge.setMiniAppFileDownloader(miniAppFileDownloader)
     }
