+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/miniapptabs/fragments/MiniAppDisplayFragment.kt
@@ -18,22 +18,13 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.rakuten.tech.mobile.miniapp.*
 import com.rakuten.tech.mobile.miniapp.ads.AdMobDisplayer
-import com.rakuten.tech.mobile.miniapp.errors.MiniAppAccessTokenError
-import com.rakuten.tech.mobile.miniapp.errors.MiniAppPointsError
 import com.rakuten.tech.mobile.miniapp.file.MiniAppCameraPermissionDispatcher
 import com.rakuten.tech.mobile.miniapp.file.MiniAppFileChooserDefault
 import com.rakuten.tech.mobile.miniapp.file.MiniAppFileDownloaderDefault
-import com.rakuten.tech.mobile.miniapp.js.MessageToContact
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import com.rakuten.tech.mobile.miniapp.js.NativeEventType
-import com.rakuten.tech.mobile.miniapp.js.chat.ChatBridgeDispatcher
-import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
-import com.rakuten.tech.mobile.miniapp.js.userinfo.Points
-import com.rakuten.tech.mobile.miniapp.js.userinfo.TokenData
-import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridgeDispatcher
 import com.rakuten.tech.mobile.miniapp.navigator.ExternalResultHandler
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
-import com.rakuten.tech.mobile.miniapp.permission.AccessTokenScope
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppDevicePermissionType
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.FragmentMiniAppDisplayBinding
@@ -53,6 +44,9 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
+private const val MINI_APP_FILE_DOWNLOAD_REQUEST_CODE = 1010
+private const val EXTERNAL_WEBVIEW_REQUEST_CODE = 100
+
 
 @Suppress("TooManyFunctions", "MagicNumber", "VariableNaming")
 class MiniAppDisplayFragment : BaseFragment(), PreloadMiniAppWindow.PreloadMiniAppLaunchListener {
@@ -67,9 +61,7 @@ class MiniAppDisplayFragment : BaseFragment(), PreloadMiniAppWindow.PreloadMiniA
     private var miniappCameraPermissionCallback: (isGranted: Boolean) -> Unit = {}
     private lateinit var sampleWebViewExternalResultHandler: ExternalResultHandler
     private lateinit var binding: FragmentMiniAppDisplayBinding
-    private val externalWebViewReqCode = 100
-    private val fileChoosingReqCode = 10101
-    private val MINI_APP_FILE_DOWNLOAD_REQUEST_CODE = 1010
+    private val FILE_CHOOSING_REQUEST_CODE = 10101
     private var appInfo: MiniAppInfo? = null
     private var isloadNew = false
     private lateinit var miniAppDisplay: MiniAppDisplay
@@ -260,7 +252,7 @@ class MiniAppDisplayFragment : BaseFragment(), PreloadMiniAppWindow.PreloadMiniA
                     sampleWebViewExternalResultHandler = externalResultHandler
                     WebViewActivity.startForResult(
                         activity, url,
-                        appId, null, externalWebViewReqCode
+                        appId, null, EXTERNAL_WEBVIEW_REQUEST_CODE
                     )
                 }
             }
@@ -294,7 +286,7 @@ class MiniAppDisplayFragment : BaseFragment(), PreloadMiniAppWindow.PreloadMiniA
             }
         }
         miniAppFileChooser = MiniAppFileChooserDefault(
-            requestCode = fileChoosingReqCode,
+            requestCode = FILE_CHOOSING_REQUEST_CODE,
             miniAppCameraPermissionDispatcher = miniAppCameraPermissionDispatcher
         )
         miniAppFileDownloader = MiniAppFileDownloaderDefault(
@@ -315,158 +307,20 @@ class MiniAppDisplayFragment : BaseFragment(), PreloadMiniAppWindow.PreloadMiniA
         miniAppFileDownloader: MiniAppFileDownloaderDefault
     ) {
         // setup MiniAppMessageBridge
-        miniAppMessageBridge = object : MiniAppMessageBridge() {
-
-            override fun getUniqueId(
-                onSuccess: (uniqueId: String) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                val errorMsg = AppSettings.instance.uniqueIdError
-                if (errorMsg.isNotEmpty()) onError(errorMsg)
-                else onSuccess(AppSettings.instance.uniqueId)
-            }
-
-            override fun getMessagingUniqueId(
-                onSuccess: (uniqueId: String) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                val errorMsg = AppSettings.instance.uniqueIdError
-                if (errorMsg.isNotEmpty()) onError(errorMsg)
-                else onSuccess("TEST-MESSAGE_UNIQUE-ID-01234")
-            }
-
-            override fun getMauid(
-                onSuccess: (mauid: String) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                val errorMsg = AppSettings.instance.mauIdError
-                if (errorMsg.isNotEmpty()) onError(errorMsg)
-                else onSuccess("TEST-MAUID-01234-56789")
-            }
-
-            override fun requestDevicePermission(
-                miniAppPermissionType: MiniAppDevicePermissionType,
-                callback: (isGranted: Boolean) -> Unit
-            ) {
-                miniappPermissionCallback = callback
-                ActivityCompat.requestPermissions(
-                    activity,
-                    AppPermission.getDevicePermissionRequest(miniAppPermissionType),
-                    AppPermission.getDeviceRequestCode(miniAppPermissionType)
-                )
-            }
-
-            override fun sendJsonToHostApp(
-                jsonStr: String,
-                onSuccess: (String) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                jsonStr.let {
-                    val requireActivity = requireActivity()
-                    val message: String
-                    if (it.isNotBlank()) {
-                        message = it
-                        onSuccess(message)
-                    } else {
-                        message =
-                            getString(R.string.error_send_json_to_host_app_please_try_again)
-                        onError(message)
-                    }
-                    if (requireActivity.isAvailable) {
-                        requireActivity.showToastMessage(message)
-                    }
-                }
-            }
+        miniAppMessageBridge = getMessageBridge(activity) { onDevicePermissionResultCallback ->
+            miniappPermissionCallback = onDevicePermissionResultCallback
         }
         miniAppMessageBridge.setAdMobDisplayer(AdMobDisplayer(activity))
         miniAppMessageBridge.allowScreenOrientation(true)
 
         // setup UserInfoBridgeDispatcher
-        val userInfoBridgeDispatcher = object : UserInfoBridgeDispatcher {
-
-            override fun getUserName(
-                onSuccess: (userName: String) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                val name = AppSettings.instance.profileName
-                if (name.isNotEmpty()) onSuccess(name)
-                else onError("User name is not found.")
-            }
-
-            override fun getProfilePhoto(
-                onSuccess: (profilePhoto: String) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                val photo = AppSettings.instance.profilePictureUrlBase64
-                if (photo.isNotEmpty()) onSuccess(photo)
-                else onError("Profile photo is not found.")
-            }
-
-            override fun getAccessToken(
-                miniAppId: String,
-                accessTokenScope: AccessTokenScope,
-                onSuccess: (tokenData: TokenData) -> Unit,
-                onError: (tokenError: MiniAppAccessTokenError) -> Unit
-            ) {
-                if (AppSettings.instance.accessTokenError != null) {
-                    onError(AppSettings.instance.accessTokenError!!)
-                } else {
-                    onSuccess(AppSettings.instance.tokenData)
-                }
-            }
-
-            override fun getContacts(
-                onSuccess: (contacts: ArrayList<Contact>) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                if (AppSettings.instance.isContactsSaved)
-                    onSuccess(AppSettings.instance.contacts)
-                else
-                    onError("There is no contact found in HostApp.")
-            }
-
-            override fun getPoints(
-                onSuccess: (points: Points) -> Unit,
-                onError: (pointsError: MiniAppPointsError) -> Unit
-            ) {
-                val points = AppSettings.instance.points
-                if (points != null) onSuccess(points)
-                else onError(MiniAppPointsError.custom("There is no points found in HostApp."))
-            }
-        }
+        val userInfoBridgeDispatcher = getUserInfoBridgeDispatcher()
         miniAppMessageBridge.setUserInfoBridgeDispatcher(userInfoBridgeDispatcher)
 
         // setup ChatBridgeDispatcher
         val chatWindow = ChatWindow(activity)
-        val chatBridgeDispatcher = object : ChatBridgeDispatcher {
-
-            override fun sendMessageToContact(
-                message: MessageToContact,
-                onSuccess: (contactId: String?) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                chatWindow.openSingleContactSelection(message, onSuccess, onError)
-            }
-
-            override fun sendMessageToContactId(
-                contactId: String,
-                message: MessageToContact,
-                onSuccess: (contactId: String?) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                chatWindow.openSpecificContactIdSelection(contactId, message, onSuccess, onError)
-            }
-
-            override fun sendMessageToMultipleContacts(
-                message: MessageToContact,
-                onSuccess: (contactIds: List<String>?) -> Unit,
-                onError: (message: String) -> Unit
-            ) {
-                chatWindow.openMultipleContactSelections(message, onSuccess, onError)
-            }
-        }
+        val chatBridgeDispatcher = getChatBridgeDispatcher(chatWindow)
         miniAppMessageBridge.setChatBridgeDispatcher(chatBridgeDispatcher)
-
         miniAppMessageBridge.setMiniAppFileDownloader(miniAppFileDownloader)
     }
 
@@ -477,7 +331,7 @@ class MiniAppDisplayFragment : BaseFragment(), PreloadMiniAppWindow.PreloadMiniA
         }
 
         when {
-            requestCode == externalWebViewReqCode && resultCode == Activity.RESULT_OK -> {
+            requestCode == EXTERNAL_WEBVIEW_REQUEST_CODE && resultCode == Activity.RESULT_OK -> {
                 data?.let { intent ->
                     val isClosedByBackPressed =
                         intent.getBooleanExtra("isClosedByBackPressed", false)
@@ -489,7 +343,7 @@ class MiniAppDisplayFragment : BaseFragment(), PreloadMiniAppWindow.PreloadMiniA
                         sampleWebViewExternalResultHandler.emitResult(intent)
                 }
             }
-            requestCode == fileChoosingReqCode && resultCode == Activity.RESULT_OK -> {
+            requestCode == FILE_CHOOSING_REQUEST_CODE && resultCode == Activity.RESULT_OK -> {
                 miniAppFileChooser.onReceivedFiles(data)
             }
             requestCode == MINI_APP_FILE_DOWNLOAD_REQUEST_CODE -> {
